+++ e2e/tests/media.e2e.js
@@ -57,6 +57,11 @@ describe('plugins', () => {
       cy.eyesCheckWindow(this.test.title + '  - plugin full width size');
     });
 
+    it('render image with link', function() {
+      cy.loadRicosEditorAndViewer('image-with-link');
+      cy.getImageLink();
+    });
+
     it('render image with loader - loading in component data', function() {
       cy.loadRicosEditorAndViewer('image-with-loader-percent');
       cy.get(`[data-hook=loader]`).should('to.be.visible');