+++ packages/editor/web/src/RichContentEditor/Toolbars/MobileBlockLinkModal.jsx
@@ -39,7 +39,7 @@ export default class MobileBlockLinkModal extends Component {
       uiSettings,
       unchangedUrl,
     } = this.props;
-    const componentLink = pubsub.getBlockData({ key: 'componentLink' });
+    const componentLink = pubsub.get('componentData')?.config?.link;
     const { url, target, rel } = componentLink || {};
     const targetBlank = target ? target === '_blank' : anchorTarget === '_blank';
     const nofollow = rel ? rel === 'nofollow' : relValue === 'nofollow';